Appease clippy by inlining dataframe.rs tests
